+++ src/completers/path.rs
@@ -171,8 +171,8 @@ fn complete_bin(sh: &shell::Shell, path: &str) -> Vec<Completion> {
         });
     }
     let builtins = vec![
-        "alias", "bg", "cd", "cinfo", "exec", "exit", "export", "fg", "history",
-        "jobs", "source", "unalias", "vox",
+        "alias", "bg", "cd", "cinfo", "exec", "exit", "export", "fg", "history", "jobs", "source",
+        "unalias", "vox",
     ];
     for item in &builtins {
         if !item.starts_with(fname) {